New optimized way to pass on EntityChanges, with Table/Rows/Row/Fields abstraction, and a map to get to its underlying fields. Stripped `old_value` and `ordinal` as its not needed by `graph-node`, Also fused `CREATE` and `UPDATE` as no distinction is made in `graph-node`.
WARN: when you create an entity, you need to write all of the non-nullable
fields on the first write _at least_. Future writes won't need to
update all fields, but the first is a constraint of the database:
if a row is created, non-nullable fields need a value.

Simplified: EventAmounts are no more, don't have a module that
does just the aggregation of such events. Rather they're computed
in the place where they're needed.