BigInt and BigDecimals to strings inside proto files.
Streamlined the Tables abstraction too, with `set` that don't
need deltas anymore. Added `set_bigint` and `set_bigdecimal`
when we want to set the typed stuff without decoding the strings
(that will assign the string as a type `Bigint` or `Bigdecimal`
without conversion).

Prefix primary keys with 0x where needed, in `graph_out` only.

TODO: fixup `flashes`, what's that anyway?

This completes the removal of `map_ticks` and `store_ticks`.